+++ multisim/src/main/java/com/hover/multisim/data/database/model/HSDKSims.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hover.multisim.db.model
+package com.hover.multisim.data.database.model
 
 import androidx.room.Entity
 import androidx.room.Index
